Make process_google_event_changes as a Celery task instead of background_task
