+++ backend/src/appointment/routes/webhooks.py
@@ -3,25 +3,22 @@ import logging
 
 import requests
 import sentry_sdk
-from fastapi import APIRouter, Depends, Request, Response, BackgroundTasks
+from fastapi import APIRouter, Depends, Request, Response
 from google.oauth2.credentials import Credentials
 from sqlalchemy.orm import Session
 
 from ..controller import auth, data, zoom
 from ..controller.apis.fxa_client import FxaClient
-from ..controller.apis.google_client import GoogleClient, SendUpdates
-from ..controller.calendar import Tools
+from ..controller.apis.google_client import GoogleClient
 from ..controller.google_watch import teardown_watch_channel
 from ..database import repo, models, schemas
-from ..database.models import MeetingLinkProviderType
-from ..defines import FALLBACK_LOCALE
 from ..dependencies.database import get_db, get_redis
 from ..dependencies.fxa import get_webhook_auth as get_webhook_auth_fxa, get_fxa_client
 from ..dependencies.google import get_google_client
 from ..dependencies.zoom import get_webhook_auth as get_webhook_auth_zoom
 from ..exceptions.account_api import AccountDeletionSubscriberFail
 from ..exceptions.fxa_api import MissingRefreshTokenException
-from ..l10n import l10n
+from ..tasks.google import process_google_event_changes
 
 router = APIRouter()
 
@@ -124,7 +121,6 @@ def zoom_deauthorization(
 @router.post('/google-calendar')
 def google_calendar_notification(
     request: Request,
-    background_tasks: BackgroundTasks,
     db: Session = Depends(get_db),
     redis=Depends(get_redis),
     google_client: GoogleClient = Depends(get_google_client),
@@ -186,226 +182,7 @@ def google_calendar_notification(
     if new_sync_token:
         repo.google_calendar_channel.update_sync_token(db, channel, new_sync_token)
 
-    background_tasks.add_task(
-        _process_google_event_changes,
-        calendar_id=calendar.id,
-        changed_events=changed_events,
-        google_client=google_client,
-        google_token=token,
-        remote_calendar_id=calendar.user,
-    )
+    process_google_event_changes.delay(calendar.id, changed_events)
 
     return success_response
 
-
-
-def _process_google_event_changes(
-    calendar_id: int,
-    changed_events: list[dict],
-    google_client: GoogleClient,
-    google_token,
-    remote_calendar_id: str,
-):
-    """Process changed events from a Google Calendar push notification.
-    Runs as a background task to avoid blocking the webhook response."""
-    from ..dependencies.database import get_engine_and_session
-
-    _, SessionLocal = get_engine_and_session()
-    db = SessionLocal()
-
-    try:
-        for event in changed_events:
-            google_event_id = event.get('id')
-            if not google_event_id:
-                continue
-
-            appointment = _find_appointment_by_external_id(db, calendar_id, google_event_id)
-            if not appointment:
-                continue
-
-            slot = appointment.slots[0] if appointment.slots else None
-            if not slot or not slot.attendee:
-                continue
-
-            # Subscriber deleted/cancelled the event from Google Calendar
-            if event.get('status') == 'cancelled':
-                _handle_event_cancelled(db, appointment, slot)
-                continue
-
-            attendees = event.get('attendees', [])
-            if not attendees:
-                continue
-
-            # Check if the subscriber (calendar owner) responded via Google Calendar
-            self_attendee = next((a for a in attendees if a.get('self')), None)
-            if self_attendee:
-                _handle_subscriber_rsvp(
-                    db, appointment, slot, self_attendee.get('responseStatus'),
-                    google_client, google_token, remote_calendar_id,
-                )
-
-            # Check if the bookee responded via Google Calendar
-            attendee_email = slot.attendee.email.lower()
-            google_attendee = next(
-                (a for a in attendees if a.get('email', '').lower() == attendee_email),
-                None,
-            )
-            if not google_attendee:
-                continue
-
-            response_status = google_attendee.get('responseStatus')
-            _handle_bookee_rsvp(
-                db, appointment, slot, response_status, google_client, google_token, remote_calendar_id
-            )
-    except Exception as e:
-        logging.error(f'[webhooks.google_calendar] Error processing event changes: {e}')
-        if sentry_sdk.is_initialized():
-            sentry_sdk.capture_exception(e)
-    finally:
-        db.close()
-
-
-def _find_appointment_by_external_id(
-    db: Session, calendar_id: int, external_id: str
-) -> models.Appointment | None:
-    """Find an appointment on the given calendar matching the Google event ID."""
-    return (
-        db.query(models.Appointment)
-        .filter(
-            models.Appointment.calendar_id == calendar_id,
-            models.Appointment.external_id == external_id,
-        )
-        .first()
-    )
-
-
-def _handle_event_cancelled(
-    db: Session,
-    appointment: models.Appointment,
-    slot: models.Slot,
-):
-    """Handle a Google Calendar event that was deleted/cancelled by the subscriber."""
-    if slot.booking_status not in (models.BookingStatus.requested, models.BookingStatus.booked):
-        return
-
-    slot_update = schemas.SlotUpdate(booking_status=models.BookingStatus.cancelled)
-    repo.slot.update(db, slot.id, slot_update)
-
-    logging.info(
-        f'[webhooks.google_calendar] Event cancelled for appointment {appointment.id}, '
-        f'slot {slot.id} marked as cancelled'
-    )
-
-
-
-def _handle_subscriber_rsvp(
-    db: Session,
-    appointment: models.Appointment,
-    slot: models.Slot,
-    response_status: str,
-    google_client: GoogleClient,
-    google_token,
-    remote_calendar_id: str,
-):
-    """React to the subscriber (calendar owner) accepting/declining via Google Calendar."""
-    if response_status == 'accepted':
-        if (
-            appointment.status != models.AppointmentStatus.opened
-            or slot.booking_status != models.BookingStatus.requested
-        ):
-            return
-
-        repo.appointment.update_status(db, appointment.id, models.AppointmentStatus.closed)
-        repo.slot.book(db, slot.id)
-
-        if appointment.external_id:
-            subscriber = appointment.calendar.owner
-            title = Tools.default_event_title(slot, subscriber)
-            repo.appointment.update_title(db, appointment.id, title)
-
-            location_url = appointment.location_url
-            if appointment.meeting_link_provider == MeetingLinkProviderType.zoom:
-                location_url = zoom.create_meeting_link(db, slot, subscriber, title) or location_url
-
-            owner_lang = subscriber.language if subscriber.language else FALLBACK_LOCALE
-            body = {'status': 'confirmed', 'summary': title}
-
-            if location_url:
-                body['location'] = location_url
-
-            description = [appointment.details or '']
-            if location_url:
-                description.append(l10n('join-online', {'url': location_url}, lang=owner_lang))
-
-            body['description'] = '\n'.join(description)
-
-            try:
-                remote_event = google_client.get_event(remote_calendar_id, appointment.external_id, google_token)
-                if remote_event and remote_event.get('attendees'):
-                    for att in remote_event['attendees']:
-                        if att.get('self'):
-                            att['responseStatus'] = 'accepted'
-                    body['attendees'] = remote_event['attendees']
-
-                google_client.patch_event(
-                    remote_calendar_id, appointment.external_id, body, google_token,
-                )
-            except Exception:
-                logging.warning('[webhooks.google_calendar] Failed to confirm event in Google')
-
-        logging.info(
-            f'[webhooks.google_calendar] Subscriber confirmed appointment {appointment.id} '
-            f'via Google Calendar, slot {slot.id} booked'
-        )
-
-    elif response_status == 'declined':
-        if slot.booking_status in (models.BookingStatus.requested, models.BookingStatus.booked):
-            slot_update = schemas.SlotUpdate(booking_status=models.BookingStatus.declined)
-            repo.slot.update(db, slot.id, slot_update)
-
-            if appointment.external_id:
-                try:
-                    google_client.delete_event(
-                        remote_calendar_id, appointment.external_id, google_token,
-                        send_updates=SendUpdates.ALL,
-                    )
-                except Exception:
-                    logging.warning('[webhooks.google_calendar] Failed to delete declined event from Google')
-
-            logging.info(
-                f'[webhooks.google_calendar] Subscriber declined appointment {appointment.id} '
-                f'via Google Calendar, slot {slot.id} marked as declined'
-            )
-
-
-def _handle_bookee_rsvp(
-    db: Session,
-    appointment: models.Appointment,
-    slot: models.Slot,
-    response_status: str,
-    google_client: GoogleClient,
-    google_token,
-    remote_calendar_id: str,
-):
-    """React to the bookee's RSVP status change from Google Calendar."""
-    if response_status == 'declined':
-        if slot.booking_status in (models.BookingStatus.requested, models.BookingStatus.booked):
-            slot_update = schemas.SlotUpdate(booking_status=models.BookingStatus.declined)
-            repo.slot.update(db, slot.id, slot_update)
-
-            if appointment.external_id:
-                try:
-                    google_client.delete_event(remote_calendar_id, appointment.external_id, google_token)
-                except Exception:
-                    logging.warning('[webhooks.google_calendar] Failed to delete declined event from Google')
-
-            logging.info(
-                f'[webhooks.google_calendar] Bookee declined appointment {appointment.id}, '
-                f'slot {slot.id} marked as declined'
-            )
-
-    elif response_status == 'accepted':
-        logging.info(
-            f'[webhooks.google_calendar] Bookee accepted appointment {appointment.id}, '
-            f'slot {slot.id}'
-        )
